Trellis.Asp: remove obsolete response verbs (post-Phase 3)
Phase 3 unified all Result-to-HTTP conversions behind a single
ToHttpResponse / ToHttpResponseAsync verb (IResult-based). This commit
deletes the seven legacy extension classes that were kept under an
[Obsolete] marker for one release cycle:

- ActionResultExtensions(Async) - MVC ToActionResult / ToCreatedAtActionResult
- HttpResultExtensions(Async)   - Minimal API ToHttpResult / ToCreatedHttpResult / ToUpdatedHttpResult
- PageActionResultExtensions     - ToPagedActionResult
- PageHttpResultExtensions       - ToPagedHttpResult
- WriteOutcomeExtensions         - WriteOutcome<T> mapping helpers

The TrellisWriteOutcomeResult IResult previously delegated to
WriteOutcomeExtensions.ToHttpResult; that switch is now inlined
directly. Likewise the Result<Page<T>> overload of ToHttpResponse
inlines PagedResponseBuilder.Build + PagedHttpResult instead of
delegating to the obsolete PageHttpResultExtensions.

The non-obsolete companion types (PagedResponse<T>, PageLink, and the
PagedHttpResult IResult wrapper) were moved from
PageHttpResultExtensions.cs into PagedResponseBuilder.cs so they
survive the deletion.

Test suites that exercised only the deleted verbs are removed; the
new API is covered by ToHttpResponseTests and the Phase 3 suites.

The CS0618/CS0619/CS0612/CS0672 obsolete-warning suppression is
removed from Trellis.Asp.csproj and Trellis.Asp.Tests.csproj: no
production code references the removed methods, so no obsolete
warnings are emitted.
