+++ Trellis.Asp/src/Trellis.Asp.csproj
@@ -6,7 +6,7 @@
     <IsAotCompatible>true</IsAotCompatible>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <TrellisApiRefName>asp</TrellisApiRefName>
-    <NoWarn>$(NoWarn);CS0618;CS0619;CS0612;CS0672</NoWarn>
+    <NoWarn>$(NoWarn)</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="Trellis.Asp.Tests" />
