+++ Trellis.Asp/tests/Trellis.Asp.Tests.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <NoWarn>$(NoWarn);CS0618;CS0619;CS0612;CS0672</NoWarn>
+    <NoWarn>$(NoWarn)</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <Using Include="Trellis.Asp" />
